terminal_python/requirements.txt
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.7.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.3.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
crcmod==1.7
cryptography==41.0.7
dbus-python==1.3.2
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
distro-info==1.7+build1
docstring_parser==0.16
docutils==0.21.2
et_xmlfile==2.0.0
fastavro==1.8.2
filelock==3.16.1
gitdb==4.0.12
GitPython==3.1.40
google-api-core==2.22.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.12.0
google-cloud-core==2.4.1
google-cloud-firestore==2.5.1
google-cloud-kms==2.19.2
google-cloud-logging==3.9.0
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.20.1
google-cloud-storage==2.12.0
google-cloud-tasks==2.14.2
google-cloud-vision==3.9.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.62.3
httplib2==0.20.4
hyperlink==21.0.0
id==1.5.0
idna==3.6
importlib_metadata==8.5.0
incremental==22.10.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyring==25.6.0
keyrings.google-artifactregistry-auth==1.1.2
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mock==5.1.0
more-itertools==10.6.0
netifaces==0.11.0
nh3==0.2.20
numpy==2.1.3
oauthlib==3.2.2
openpyxl==3.1.0
opentelemetry-api==1.29.0
packaging==24.2
pandas==2.2.2
pillow==10.2.0
pipdeptree==2.25.0
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.25.0
protobuf==4.25.6
pyarrow==16.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.45.3
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyproject_hooks==1.2.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
readme_renderer==44.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rsa==4.9
SecretStorage==3.3.3
service-identity==24.1.0
setuptools==68.1.2
shtab==1.7.1
six==1.16.0
smmap==5.0.2
systemd-python==235
termcolor==2.5.0
tldr==3.3.0
twine==6.1.0
Twisted==24.3.0
typing_extensions==4.10.0
tzdata==2024.2
ubuntu-pro-client==8001
unattended-upgrades==0.1
Unidecode==1.3.6
urllib3==2.0.7
virtualenv==20.27.1
wadllib==1.3.6
wheel==0.42.0
wrapt==1.17.2
zipp==3.21.0
zope.interface==6.1

